It makes it possible to load asynchronously
Sometimes it is necessary to add scripts dynamically. If you enable async option (true) scripts to load asynchronously. If false, pjax will operate as usual.
Why do it? Scripts are loaded on the fly, and client scripts waiting until all the necessary scripts to connect to and then work out fine.